requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.2
sqlalchemy[asyncio]~=2.0.41
asyncpg
bcrypt
python-jose[cryptography]~=3.4.0
pydantic~=2.11.4
alembic~=1.16.0
psycopg2
passlib~=1.7.4
python-multipart
PySide6~=6.9.0
requests~=2.32.3
python-dotenv~=1.1.0
openpyxl
pyinstaller